+++ src/app/[submitted]/page.tsx
@@ -5,7 +5,7 @@ import { Stack, Typography, Box, Button } from "@mui/material"
 import { useLazyQuery } from "@apollo/client"
 import { client } from "../client"
 import { ElementFilterState, SequenceFilterState, GeneFilterState, CCREs, InputRegions } from "../types"
-import Filters, { initialElementFilterState, initialGeneFilterState, initialSequenceFilterState } from "./tempFilters"
+import Filters, { initialElementFilterState, initialGeneFilterState, initialSequenceFilterState } from "./Filters"
 import { BED_INTERSECT_QUERY } from "../queries"
 import { decodeRegions } from "../_utility/coding"
 import Link from "next/link"
