+++ src/app/[submitted]/tables/genes/GeneTable.tsx
