+++ src/app/[submitted]/tables/sequence/MotifModal.tsx
