+++ src/app/page.tsx
@@ -2,7 +2,7 @@
 "use client";
 import { Box, Stack, Typography } from "@mui/material";
 import React from "react";
-import ArgoUpload from "./components/tempArgoUpload";
+import ArgoUpload from "./components/ArgoUpload";
 
 export default function Home() {
   return (
